+++ src/pushnotification/service.hh
@@ -38,6 +38,11 @@ namespace flexisip::pushnotification {
 class Client;
 class GenericHttpRequest;
 
+struct ApnsServers {
+	std::string prod;
+	std::string dev;
+};
+
 class Service {
 public:
 	Service(const std::shared_ptr<sofiasip::SuRoot>& root, unsigned maxQueueSize);
@@ -61,7 +66,7 @@ public:
 	std::shared_ptr<Request> makeRequest(PushType pType, const std::shared_ptr<const PushInfo>& pInfo);
 	void sendPush(const std::shared_ptr<Request>& pn);
 	void setupGenericClient(const sofiasip::Url& url, Method method, Protocol protocol);
-	void setupiOSClient(const std::string& certDir, const std::string& caFile);
+	void setupiOSClient(const std::string& certDir, const std::string& caFile, const ApnsServers& apnsServers);
 	void setupFirebaseClients(const GenericStruct* pushConfig);
 	void addFirebaseClient(const std::string& appId, const std::string& apiKey = "");
 	void addFirebaseV1Client(const std::string& appId,
@@ -78,6 +83,7 @@ public:
 	bool isIdle() const noexcept;
 
 	static const std::string sGenericClientName;
+	static const ApnsServers kDefaultApnsServers;
 
 private:
 	static const std::string sFallbackClientKey;
@@ -94,6 +100,7 @@ private:
 	std::map<std::string, std::shared_ptr<Client>> mClients{};
 	std::string mWindowsPhonePackageSID{};
 	std::string mWindowsPhoneApplicationSecret{};
+	ApnsServers mApnsServers{};
 	std::map<std::filesystem::path, std::filesystem::path> mAppleCertDirs{};
 	StatCounter64* mCountFailed{nullptr};
 	StatCounter64* mCountSent{nullptr};