+++ tester/tests/pushnotification/push-notification-tester.cc
@@ -119,9 +119,8 @@ static void startApplePushTest(PushType pType,
                                const string& responseBody,
                                Request::State expectedFinalState,
                                bool timeout = false) {
-	AppleClient::APN_DEV_ADDRESS = "127.0.0.1";
 	AppleClient::APN_PORT = "3000";
-	AppleClient appleClient{*root, "", bcTesterRes("cert/apple.test.dev.pem"), "apple.test.dev"};
+	AppleClient appleClient{*root, "", bcTesterRes("cert/apple.test.dev.pem"), "127.0.0.1"};
 	appleClient.enableInsecureTestMode();
 
 	auto request = make_shared<AppleRequest>(pType, pushInfo);
@@ -598,9 +597,8 @@ static void applePushTestConnectErrorAndReconnect() {
 \})json"};
 
 	// We first send a request with mock off, leading to TLS connection error.
-	AppleClient::APN_DEV_ADDRESS = "127.0.0.1";
 	AppleClient::APN_PORT = "3000";
-	AppleClient appleClient{*root, "", bcTesterRes("cert/apple.test.dev.pem"), "apple.test.dev"};
+	AppleClient appleClient{*root, "", bcTesterRes("cert/apple.test.dev.pem"), "127.0.0.1"};
 	appleClient.enableInsecureTestMode();
 
 	auto request = make_shared<AppleRequest>(PushType::Message, pushInfo);